+++ src/filters.py
@@ -6,121 +6,6 @@ from langchain_community.embeddings import OpenAIEmbeddings
 from qdrant_client import QdrantClient
 
 
-def initialize_session_state():
-    """Initialize session state variables for API keys and URLs with environment defaults."""
-    # Get environment variables
-    env_openai_key = os.getenv("OPENAI_API_KEY", "")
-    env_qdrant_url = os.getenv("QDRANT_URL", "")
-    env_qdrant_key = os.getenv("QDRANT_KEY", "")
-    
-    # Initialize session state with environment variables
-    if 'openai_api_key' not in st.session_state:
-        st.session_state.openai_api_key = env_openai_key
-    if 'qdrant_url' not in st.session_state:
-        st.session_state.qdrant_url = env_qdrant_url
-    if 'qdrant_api_key' not in st.session_state:
-        st.session_state.qdrant_api_key = env_qdrant_key
-    if 'using_env_vars' not in st.session_state:
-        st.session_state.using_env_vars = True
-    if 'must_filters' not in st.session_state:
-        st.session_state.must_filters = []
-    if 'must_not_filters' not in st.session_state:
-        st.session_state.must_not_filters = []
-    if 'should_filters' not in st.session_state:
-        st.session_state.should_filters = []
-    if 'embedding_type' not in st.session_state:
-        st.session_state.embedding_type = 'openai'
-
-def create_sidebar():
-    """Create sidebar for API configuration with environment variable defaults."""
-    with st.sidebar:
-        st.header("Configuration Settings")
-        
-        # Create expander for API settings
-        with st.expander("API Configuration", expanded=False):
-            # Embedding type selector
-            st.session_state.embedding_type = st.radio(
-                "Embedding Type",
-                options=['openai', 'fastembed'],
-                index=0 if st.session_state.embedding_type == 'openai' else 1,
-                help="Select embedding provider"
-            )
-            
-            # Toggle for using environment variables
-            st.session_state.using_env_vars = st.checkbox(
-                "Use Environment Variables",
-                value=st.session_state.using_env_vars,
-                help="Toggle between environment variables and custom settings"
-            )
-            
-            if st.session_state.using_env_vars:
-                # Display current environment variable values
-                st.info("Using environment variables")
-                st.write("Environment Variables Status:")
-                if st.session_state.embedding_type == 'openai':
-                    st.write("- OpenAI API Key:", "✅ Set" if os.getenv("OPENAI_API_KEY") else "❌ Not Set")
-                st.write("- Qdrant URL:", "✅ Set" if os.getenv("QDRANT_URL") else "❌ Not Set")
-                st.write("- Qdrant API Key:", "✅ Set" if os.getenv("QDRANT_KEY") else "❌ Not Set")
-                
-                # Reset to environment variables
-                st.session_state.qdrant_url = os.getenv("QDRANT_URL", "")
-                st.session_state.qdrant_api_key = os.getenv("QDRANT_KEY", "")
-                if st.session_state.embedding_type == 'openai':
-                    st.session_state.openai_api_key = os.getenv("OPENAI_API_KEY", "")
-            else:
-                # Custom configuration inputs
-                if st.session_state.embedding_type == 'openai':
-                    openai_key = st.text_input(
-                        "OpenAI API Key",
-                        value=st.session_state.openai_api_key,
-                        type="password",
-                        help="Enter your OpenAI API key"
-                    )
-                    if openai_key:
-                        st.session_state.openai_api_key = openai_key
-                
-                qdrant_url = st.text_input(
-                    "Qdrant URL",
-                    value=st.session_state.qdrant_url,
-                    help="Enter your Qdrant server URL"
-                )
-                if qdrant_url:
-                    st.session_state.qdrant_url = qdrant_url
-                
-                qdrant_key = st.text_input(
-                    "Qdrant API Key",
-                    value=st.session_state.qdrant_api_key,
-                    type="password",
-                    help="Enter your Qdrant API key"
-                )
-                if qdrant_key:
-                    st.session_state.qdrant_api_key = qdrant_key
-                
-                # Save button (only shown when using custom settings)
-                if st.button("Save Configuration"):
-                    st.success("Custom configuration saved successfully!")
-        
-        # Show current active configuration
-        st.write("Active Configuration:")
-        st.write("Using:", "Environment Variables" if st.session_state.using_env_vars else "Custom Settings")
-        st.write("Embedding Provider:", st.session_state.embedding_type.upper())
-        if st.session_state.embedding_type == 'openai':
-            st.write("- OpenAI API Key:", "✅ Set" if st.session_state.openai_api_key else "❌ Not Set")
-        st.write("- Qdrant URL:", "✅ Set" if st.session_state.qdrant_url else "❌ Not Set")
-        st.write("- Qdrant API Key:", "✅ Set" if st.session_state.qdrant_api_key else "❌ Not Set")
-
-def get_embeddings(query: str):
-    """Get embeddings based on selected provider."""
-    if st.session_state.embedding_type == 'openai':
-        embedding = OpenAIEmbeddings(api_key=st.session_state.openai_api_key)
-        return embedding.embed_query(query)
-    else:
-        # Use Qdrant's built-in FastEmbed
-        return {
-            "name": "fastembed",
-            "query": query
-        }
-    
 def create_filter_condition(
     filter_type: str, key: str, operator: str, value: Any
 ) -> Dict:
@@ -336,66 +221,3 @@ def create_filter_interface(metadata_keys: List[str]) -> Dict[str, List[Dict]]:
         st.json(filter_clause)
 
     return filter_clause
-
-def main():
-    initialize_session_state()
-    create_sidebar()
-    st.title("Advanced Qdrant Query with Filters")
-
-    client = QdrantClient(url=st.session_state.qdrant_url, api_key=st.session_state.qdrant_api_key)
-    
-    query = st.text_input("Enter your query:")
-    limit = st.number_input("Limit", min_value=1, max_value=20, value=5)
-
-    # Get collections
-    collections = client.get_collections()
-    collection_names = [collection.name for collection in collections.collections]
-    collection_name = st.selectbox("Select Collection", collection_names)
-
-    # Get metadata keys from collection
-    record, pointId = client.scroll(collection_name=collection_name, limit=1)
-    metadata_keys = list(record[0].payload.keys())
-
-    # Create filter interface
-    filter_clause = create_filter_interface(metadata_keys)
-
-    # Query button and results
-    if st.button("Query Qdrant"):
-        try:
-            if st.session_state.embedding_type == 'openai':
-                # OpenAI embedding
-                query_embedding = get_embeddings(query)
-                results = client.search(
-                    collection_name=collection_name,
-                    query_vector=query_embedding,
-                    limit=limit,
-                    query_filter=filter_clause,
-                )
-            else:
-                # FastEmbed
-                results = client.search(
-                    collection_name=collection_name,
-                    query_text=query,  # Send raw text for server-side embedding
-                    limit=limit,
-                    query_filter=filter_clause,
-                    append_payload=True,  # Make sure we get payloads back
-                )
-
-            if not results:
-                st.write("No results found")
-
-            for i, hit in enumerate(results):
-                with st.expander(f"Result {i+1}", expanded=True):
-                    # extract keys from hit.payload
-                    keys = list(hit.payload.keys())
-                    payload_dict = {key: hit.payload.get(key, "N/A") for key in keys}
-                    
-                    st.json(payload_dict)
-                    
-        except Exception as e:
-            st.error(f"Error during search: {str(e)}")
-            if st.session_state.embedding_type == 'fastembed':
-                st.info("Note: Make sure FastEmbed is properly configured on your Qdrant server")
-
-if __name__ == "__main__":
-    main()
